The only way to fix the svg bug is to write around it
So we kept running into svg bugs.
The bug report:
http://bugs.shoebot.net/issue25

We did not have time to wait for it to be fixed—
and due to the tight coupling with Cairo Shoebot
is too complicated for us to fix.

So for now we wrote around it, using another library,
svgwrite, to generate the SVG. Since we still
use Shoebot to calculate the textwidth, our
solution is quite convoluted now.
But it will have to do for now!

In this version I also added proper sizing according
to the size parameters given in the interface.